SAA-1143 Add a filter to remove expired cancellation appointments from the attendance list.
