+++ server/services/unlockListService.ts
@@ -1,46 +1,10 @@
-import { subDays, subMonths, subWeeks } from 'date-fns'
 import { SubLocationCellPattern, UnlockListItem } from '../@types/activities'
 import PrisonerSearchApiClient from '../data/prisonerSearchApiClient'
 import ActivitiesApiClient from '../data/activitiesApiClient'
 import { ServiceUser } from '../@types/express'
-import { scheduledEventSort, toDate, toDateString } from '../utils/utils'
+import { scheduledEventSort, toDateString } from '../utils/utils'
 import AlertsFilterService from './alertsFilterService'
-import { ScheduledEvent } from '../@types/activitiesAPI/types'
-import { AppointmentFrequency } from '../@types/appointments'
-
-export function applyCancellationDisplayRule(app: ScheduledEvent): boolean {
-  let showAppointment = true
-  const beginingOfToday = new Date().setHours(0, 0, 0, 0)
-  if (app.cancelled) {
-    if (
-      app.appointmentSeriesFrequency === AppointmentFrequency.DAILY &&
-      toDate(app.appointmentSeriesCancellationStartDate) < subDays(beginingOfToday, 2)
-    ) {
-      showAppointment = false
-    } else if (
-      app.appointmentSeriesFrequency === AppointmentFrequency.WEEKDAY &&
-      toDate(app.appointmentSeriesCancellationStartDate) < subDays(beginingOfToday, 4)
-    ) {
-      showAppointment = false
-    } else if (
-      app.appointmentSeriesFrequency === AppointmentFrequency.WEEKLY &&
-      toDate(app.appointmentSeriesCancellationStartDate) < subWeeks(beginingOfToday, 2)
-    ) {
-      showAppointment = false
-    } else if (
-      app.appointmentSeriesFrequency === AppointmentFrequency.FORTNIGHTLY &&
-      toDate(app.appointmentSeriesCancellationStartDate) < subWeeks(beginingOfToday, 4)
-    ) {
-      showAppointment = false
-    } else if (
-      app.appointmentSeriesFrequency === AppointmentFrequency.MONTHLY &&
-      toDate(app.appointmentSeriesCancellationStartDate) < subMonths(beginingOfToday, 2)
-    ) {
-      showAppointment = false
-    }
-  }
-  return showAppointment
-}
+import applyCancellationDisplayRule from '../utils/applyCancellationDisplayRule'
 
 export default class UnlockListService {
   constructor(
